Create js_7
The weight is stored in the 'weight' variable. Write a recommendation to the 'recommendation' variable as a string. 
If the weight is up to 4 kg (excluding this value), the recommendation is 'Time to eat'. If the weight is from 4 kg inclusive to 5.5 kg inclusive - 'The weight is normal.' If the weight is more than 5.5 kg - 'Time to workout'. Print the recommendation to the console.