JS_8
Write a function that calculates a student's final grade based on two parameters: the exam grade and the number of their completed projects. 
The function must take two arguments: exam - the score for the exam (from 0 to 100); projects - the number of the completed projects (from 0 and above); This function should return a number (final grade). 

1/ 100, if the exam score is more than 90 or the number of completed projects is more than 10.
2/ 90, if the exam score is more than 75 or the number of completed projects is at least 5.
3/ 75, if the exam score is more than 50 or the number of completed projects is at least 2.
4/ 0, otherwise

   finalGrade(100, 12) => 100
   finalGrade(99, 0) => 100
   finalGrade(10, 15) => 100
   finalGrade(85, 5) => 90
   finalGrade(55, 3) => 75
   finalGrade(55, 0) => 75
   finalGrade(20, 2) => 75