JS_11
In this exercise, you should write a program that takes two arrays of integers as parameters. It is necessary to find a product of the elements of each array, and then find the difference between the arrays' products, regardless of which one is larger. If you are able to write the solution in one line, you will be great.

   example([3, 2, 5], [1, 4, 4]) => 14
   example([9, 7, 2], [5, 2, 2]) => 106
   example([11, 2, 5], [1, 10, 8]) => 30
   example([4, 4, 7], [3, 9, 3]) => 31
   example([15, 20, 25], [10, 30, 25]) => 0