JS_15
Write a function that returns the difference between the largest and the smallest values in the list received as an argument to the function. The array that the function receives as an argument can contain positive and negative numbers. If the array is empty or has one value, return zero. Initially, the array will be unsorted. 
Example: 
maxDiff ([1, 2, 3, -4]); // will return 7 because: 3 - (-4) == 7