JS_17
A trick how to determine whether a number is divisible by 3. Find a sum of all the integers of a number and divide the total by 3. If there is no remainder of the sum, then the original number is also divisible by 3. Determine if the number represented by a string is divisible by 3. 
Note: Try to avoid using the % (modulo) operator. 

Sample code:
   divisibleByThree ('123') => true
   divisibleByThree ('19254') => true
   divisibleByThree ('88 ') => false
   divisibleByThree ('1') => false